+++ app/src/main/java/dev/yokai/domain/download/DownloadPreferences.kt
@@ -1,0 +1,9 @@
+package dev.yokai.domain.download
+
+import eu.kanade.tachiyomi.core.preference.PreferenceStore
+
+class DownloadPreferences(
+    private val preferenceStore: PreferenceStore,
+) {
+    fun downloadWithId() = preferenceStore.getBoolean("download_with_id", false)
+}